Check in the build
Because my mac (apple M3 chip) is having a lot of troubel building this, particularly
around gulp-sass, which I'd prefer not to deal with. I believe it'll require installing
dart-sass, which I'm sure is doablle, but I just don't want to do. So the idea is to build
this on linux and use it on mac.